Initial work on the from-Elm cheatsheet
Starting from the Elixir cheatsheet at Louis' recommendation. I've added
an overview to give a broad comparison before proceeding to edit the
document line by line to shape to Elm instead.

It is not completed. I have removed the unchanged parts for the moment
though the entries in the table of contents still correspond to them.