Reorganise for Elm centric viewpoint
The original version of this document started with the Elixir cheatsheet
and converted over to Elm section by section. But that meant that the
skeleton was based on an Elixir dev's knowledge of Elixir. Here we
change the outline to topics that are more familiar to an Elm dev and
leave 'TODO' comments for any section that we've yet to complete.

This hopefully looks like a solid list of topics that an Elm dev is
familiar with and would like to see 'like for like' comparisons of. We
include an 'Other concepts' section at the bottom for Atoms which seem
vaguely notable but have no Elm equivalent and possibly other topics
like OTP and anything that comes into Gleam from the Erlang foundations.