Improve the 'lists' section
By removing the slightly confusing comment about pattern matching on
Elm lists as we don't go on to demostrate it. As pointed out by
@jfmengels.

We talk about the operators more and link to the gleam list docs.

I think we should have a section on case-statements and pattern matching
but it shouldn't be here.