Switch iodata to string_builder
Per https://github.com/gleam-lang/gleam/discussions/1618#discussioncomment-2708606

I originally made this change through github UI, but I realized that
wouldn't build the changes. I cloned locally and ran `make build`, but
that created changes to many files.

So either my system is not configured for generating the docs correctly
or the docs are intentionally not built up-to-date. Either way, I
figured it best to just push the markdown change.